Make GM realization data source part of the policy one
